% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_sequence_alignments.R
\name{alignToGenome}
\alias{alignToGenome}
\title{Align queries against intervals extracted from a genome}
\usage{
alignToGenome(query.seqs, genome, gr, ...)
}
\arguments{
\item{query.seqs}{DNAStringSet object}

\item{genome}{DNAStringSet object}

\item{gr}{GRanges object}

\item{...}{Arguments passed to seqFromGenomicRange}
}
\value{
desc
}
\description{
Align queries against intervals extracted from a genome
}
