% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{class}
\name{config-class}
\alias{config-class}
\alias{config}
\title{The top configuration class}
\description{
The main thing you will need to change is the input.
}
\section{Slots}{

\describe{
\item{\code{input}}{Paths to inputs and the focal species name}

\item{\code{synder}}{Parameters for synder}

\item{\code{alignment}}{Alignment configurations}
}}

