% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{config_alignment-class}
\alias{config_alignment-class}
\alias{config_alignment}
\title{Alignment settings and thresholds}
\description{
For details on inputs see main package documentation
}
\section{Slots}{

\describe{
\item{\code{thresholds}}{p-value thresholds for hit significanc}

\item{\code{simulation}}{simulation parameters}

\item{\code{dna2dna_maxspace}}{largest size of string to compare}

\item{\code{indel_threshold}}{I don't remember what this does}
}}

