% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_orffaa}
\alias{derive_orffaa}
\title{Derive ORF protein sequences from genome and ORF ranges}
\usage{
derive_orffaa(orfgff, dna)
}
\arguments{
\item{orfgff}{GenomicRanges object}

\item{dna}{DNAStringSet genome}
}
\value{
AAStringSet object
}
\description{
Derive ORF protein sequences from genome and ORF ranges
}
