% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_derive.R
\name{derive_orfgff}
\alias{derive_orfgff}
\title{Derive mono-exonic ORF intervals from a genome}
\usage{
derive_orfgff(dna)
}
\arguments{
\item{dna}{DNAStringSet genome}
}
\value{
GenomicRanges object
}
\description{
Derive mono-exonic ORF intervals from a genome
}
