% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_derive.R
\name{derive_transcript_ORFs}
\alias{derive_transcript_ORFs}
\title{Derive ORFs from transcripts}
\usage{
derive_transcript_ORFs(dna, con)
}
\arguments{
\item{con}{fagin config object}

\item{x}{DNAStringSet transcriptome}
}
\value{
GenomicRanges object
}
\description{
Derive ORFs from transcripts
}
