% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_transorf}
\alias{derive_transorf}
\title{Derive ORF intervals from a transcriptome}
\usage{
derive_transorf(trans)
}
\arguments{
\item{trans}{DNAStringSet genome}
}
\value{
GenomicRanges object
}
\description{
Derive ORF intervals from a transcriptome
}
