% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_transorffaa}
\alias{derive_transorffaa}
\title{Derive ORF intervals from a transcriptome}
\usage{
derive_transorffaa(trans, transorfgff)
}
\arguments{
\item{trans}{DNAStringSet genome}

\item{transorfgff}{GRanges locations of ORFs on mRNAs}
}
\value{
GenomicRanges object
}
\description{
Derive ORF intervals from a transcriptome
}
