% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_transorfgff}
\alias{derive_transorfgff}
\title{Derive ORF intervals from a transcriptome}
\usage{
derive_transorfgff(trans)
}
\arguments{
\item{trans}{DNAStringSet genome}
}
\value{
GenomicRanges object
}
\description{
Derive ORF intervals from a transcriptome
}
