% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{dnaregex}
\alias{dnaregex}
\title{Internal function for extracting intervals from pattern matches on DNA}
\usage{
dnaregex(x, pattern, ...)
}
\arguments{
\item{x}{DNAStringSet}

\item{pattern}{character regular expression}

\item{...}{extra arguments for gregexpr}
}
\value{
GRanges object
}
\description{
Internal function for extracting intervals from pattern matches on DNA
}
