% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{effector}
\alias{describe_faa}
\alias{describe_feat_cds}
\alias{describe_feat_gen}
\alias{describe_feat_ind}
\alias{describe_feat_nst}
\alias{describe_feat_nuc}
\alias{describe_feat_orf}
\alias{describe_feat_res}
\alias{describe_feat_rna}
\alias{describe_feat_scr}
\alias{describe_feat_tec}
\alias{describe_feat_trn}
\alias{describe_feat_una}
\alias{describe_feat_unknown_unknowns}
\alias{describe_features}
\alias{describe_fna}
\alias{describe_frn}
\alias{describe_gff}
\alias{describe_species_tree}
\alias{describe_synder}
\alias{describe_synmap}
\alias{effector}
\title{Effector functions}
\usage{
describe_synmap(a, b)

describe_synder(a, b)

describe_gff(a, b)

describe_fna(a, b)

describe_frn(a, b)

describe_faa(a, b)

describe_species_tree(a, b)

describe_features(a, b)

describe_feat_gen(a, b)

describe_feat_trn(a, b)

describe_feat_orf(a, b)

describe_feat_nuc(a, b)

describe_feat_cds(a, b)

describe_feat_rna(a, b)

describe_feat_tec(a, b)

describe_feat_una(a, b)

describe_feat_ind(a, b)

describe_feat_nst(a, b)

describe_feat_res(a, b)

describe_feat_scr(a, b)

describe_feat_unknown_unknowns(a, b)
}
\arguments{
\item{a, b}{input and output data for parent function}
}
\description{
Effector functions
}

