% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{extractWithComplements}
\alias{extractWithComplements}
\title{Extract DNA intervals, reverse complementing them if they are minus sense}
\usage{
extractWithComplements(dna, gff)
}
\arguments{
\item{dna}{DNA template}

\item{gff}{Intervals to extract}
}
\value{
DNAStringSet
}
\description{
Extract DNA intervals, reverse complementing them if they are minus sense
}
