% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{faa_summary-class}
\alias{faa_summary-class}
\alias{faa_summary}
\title{Summary of a set of protein sequences}
\description{
Summary of a set of protein sequences
}
\section{Slots}{

\describe{
\item{\code{class_composition}}{numeric}
}}

