% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filenaming.R
\name{fagin_naming}
\alias{fagin_naming}
\alias{get_genome_filename}
\alias{get_gff_filename}
\alias{get_synmap_filename}
\title{Naming stuff}
\usage{
get_genome_filename(species_name, dir)

get_gff_filename(species_name, dir)

get_synmap_filename(focal_name, target_name, dir)
}
\arguments{
\item{species_name}{The base name for a species (spaces allowed)}

\item{dir}{The base directory}

\item{focal_name}{The name of the focal species}

\item{target_name}{The name of the target species}
}
\description{
Spaces in species names will be automatically converted to underscores.
}
