% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{fagin_printer}
\alias{fagin_printer}
\alias{print.config_alignment_thresholds}
\alias{print.config_alignment_simulation}
\alias{print.config_alignment}
\alias{print.config_input}
\alias{print.config_synder}
\alias{print.config}
\title{Special printers}
\usage{
\method{print}{config_alignment_thresholds}(x, ...)

\method{print}{config_alignment_simulation}(x, ...)

\method{print}{config_alignment}(x, ...)

\method{print}{config_input}(x, ...)

\method{print}{config_synder}(x, ...)

\method{print}{config}(x, ...)
}
\arguments{
\item{x}{An object to print}

\item{...}{Additional arguments going to God knows where}
}
\description{
Special printers
}
