% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{fagin_summary}
\alias{fagin_summary}
\alias{summarize_numeric}
\alias{summarize_faa}
\alias{summarize_fna}
\alias{summarize_gff}
\alias{summarize_nstring}
\alias{summarize_syn}
\title{Summarize functions}
\usage{
summarize_numeric(x)

summarize_faa(x)

summarize_fna(x)

summarize_gff(x)

summarize_nstring(x)

summarize_syn(x)
}
\arguments{
\item{x}{An object to summarized}
}
\description{
Summarize functions
}
