% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{feature_builders}
\alias{feature_builders}
\alias{find_cds}
\alias{find_gen}
\alias{find_ind}
\alias{find_nst}
\alias{find_nuc}
\alias{find_orf}
\alias{find_res}
\alias{find_rna}
\alias{find_scr}
\alias{find_tec}
\alias{find_trn}
\alias{find_una}
\alias{find_unknown_unknowns}
\title{Feature vector generating functions}
\usage{
find_gen(x)

find_trn(x)

find_orf(x)

find_nuc(x)

find_cds(x)

find_rna(x)

find_tec(x)

find_una(x)

find_ind(x)

find_nst(x)

find_res(x)

find_scr(x)

find_unknown_unknowns(x)
}
\arguments{
\item{x}{input data}
}
\value{
logical vector
}
\description{
Is AA similar to a known gene
gen
primary: ORFic
secondary: O1
label: ORFic - known protein
}
\details{
Is AA similar to any ORF on known mRNA?
trn
divisor: Is AA similar to any ORF on known mRNA?
primary: ORFic
secondary: O2
label: ORFic - transcribed ORF

Is AA similar to any ORF anywhere?
orf
divisor: Is AA similar to any ORF anywhere?
primary: ORFic
secondary: O3
label: ORFic - unknown ORF

Is DNA similar to anything?
nuc

Does search interval overlap a CDS?
cds
divisor: Does search interval overlap a CDS?
N1:
primary: Non-ORFic
secondary: N1
label: SI overlaps CDS

Does search interval overlap an exon?
rna
divisor: Does search interval overlap an exon?
N2:
primary: Non-ORFic
secondary: N2
label: SI overlaps exon
Was the entry skipped for technical reasons?
tec:
U7:
primary: Unknown
secondary: U7
label: skipped

Query maps off target scaffold
una:
divisor: Query maps off target scaffold
U1:
primary: Unknown
secondary: U1
label: unassembled

Query maps to zero-length target interval
ind:
divisor: Query maps to zero-length target interval
U2:
primary: Unknown
secondary: U2
label: possible indel

Query maps to N-string
nst:
divisor: Query maps to N-string
U3:
primary: Unknown
secondary: U3
label: possibly in unknown region

Query maps to target interval smaller than self
res:
divisor: Query maps to target interval smaller than self
U4:
primary: Unknown
secondary: U4
label: possibly resized

Query maps inbetween contiguous block
scr:
divisor: Query maps inbetween contiguous block
U5:
primary: Unknown
secondary: U5
label: scrambled synteny

Good syntenic match, no homology
U6:
primary: Unknown
secondary: U6
}

