% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{final_protein_residue_counts}
\alias{final_protein_residue_counts}
\title{Get a table listing the last protein residue proportions for each species}
\usage{
final_protein_residue_counts(x)
}
\description{
Get a table listing the last protein residue proportions for each species
}
