% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{get_annotated_mRNA_orfs}
\alias{get_annotated_mRNA_orfs}
\title{Get all ORFs from annotated mRNAs}
\usage{
get_annotated_mRNA_orfs(x)
}
\arguments{
\item{x}{Genome}
}
\value{
FFN
}
\description{
Get all ORFs from annotated mRNAs
}

