% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{get_genome_lengths}
\alias{get_genome_lengths}
\title{Calculate the lengths of each scaffold in a given genome}
\usage{
get_genome_lengths(x)
}
\arguments{
\item{x}{Genome}
}
\value{
([Conid], [Integer]) data frame
}
\description{
Calculate the lengths of each scaffold in a given genome
}

