% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_sequence_alignments.R
\name{get_prot2prot}
\alias{get_prot2prot}
\title{CDS feature wrapper for AA_aln}
\usage{
get_prot2prot(query, target, features, ...)
}
\description{
Inputs:
map - [ query | target ] where target is a CDS id
query - full query dataset
target - full target dataset
}
