% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{get_proteins}
\alias{get_proteins}
\title{Get proteins from all predicted genomic models}
\usage{
get_proteins(x)
}
\arguments{
\item{x}{([Genome], [GFF])}
}
\value{
[Protein]
}
\description{
Get proteins from all predicted genomic models
}

