% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{get_search_intervals}
\alias{get_search_intervals}
\title{Predict search intervals from synteny map}
\usage{
get_search_intervals(x)
}
\arguments{
\item{x}{(Synmap, Seqids)}
}
\value{
SearchIntervals
}
\description{
Predict search intervals from synteny map
}

