% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_species_from_tree}
\alias{get_species_from_tree}
\title{Get species from tree}
\usage{
get_species_from_tree(tree)
}
\arguments{
\item{tree}{phylo object representing all under inspection}
}
\value{
character vector of unique species
}
\description{
Get species from tree
}
