% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{get_species_phylogenetic_order}
\alias{get_species_phylogenetic_order}
\title{Get the phylogenetic order of species (mostly useful for plotting)}
\usage{
get_species_phylogenetic_order(con)
}
\arguments{
\item{con}{Config object}
}
\description{
Get the phylogenetic order of species (mostly useful for plotting)
}
