% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{get_synder_count_table}
\alias{get_synder_count_table}
\title{Get summary table of synder search intervals per gene counts}
\usage{
get_synder_count_table(m, genes = NULL)
}
\arguments{
\item{m}{Rmonad object}
}
\description{
Get summary table of synder search intervals per gene counts
}
