% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{get_synder_nsi}
\alias{get_synder_nsi}
\title{Get tables of number of search intervals}
\usage{
get_synder_nsi(m)
}
\arguments{
\item{m}{Rmonad object}
}
\description{
Get tables of number of search intervals
}
