% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{get_transcript_orfs}
\alias{get_transcript_orfs}
\title{Get all ORFs from transcripts}
\usage{
get_transcript_orfs(x)
}
\arguments{
\item{x}{Genome}
}
\value{
FFN
}
\description{
Get all ORFs from transcripts
}

