% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{getSpeciesPhylogeneticOrder}
\alias{getSpeciesPhylogeneticOrder}
\title{Get the phylogenetic order of species (mostly useful for plotting)}
\usage{
getSpeciesPhylogeneticOrder(con)
}
\arguments{
\item{con}{Config object}
}
\description{
Get the phylogenetic order of species (mostly useful for plotting)
}
