% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{gff_summary-class}
\alias{gff_summary-class}
\alias{gff_summary}
\title{Summary of a GFF file}
\description{
Summary of a GFF file
}
\section{Slots}{

\describe{
\item{\code{n_features}}{integer}

\item{\code{mRNA_per_model}}{numeric_summary}

\item{\code{exon_per_mRNA}}{numeric_summary}

\item{\code{cds_per_mRNA}}{numeric_summary}

\item{\code{exon_per_cds}}{numeric_summary}
}}

