% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{initial_protein_residue_counts}
\alias{initial_protein_residue_counts}
\title{Get a table listing the first protein residue proportions for each species}
\usage{
initial_protein_residue_counts(x)
}
\description{
Get a table listing the first protein residue proportions for each species
}
