% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_config}
\alias{load_config}
\title{Load the configuration file for the Fagin workflow}
\usage{
load_config(filename = "inst/etc/config-template.yml")
}
\arguments{
\item{filename}{YAML configuration file}
}
\description{
Load the configuration file for the Fagin workflow
}
