% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{load_data}
\alias{load_data}
\title{Derive secondary data from the minimal required inputs}
\usage{
load_data(con)
}
\arguments{
\item{con}{The config object that provides paths to required data}
}
\value{
derived_input object
}
\description{
Derive secondary data from the minimal required inputs
}
