% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{load_genome}
\alias{load_genome}
\title{Load a species DNA genome}
\usage{
load_genome(name, genome)
}
\arguments{
\item{name}{species name}

\item{genome}{a genomic fasta file name}
}
\description{
Load a species DNA genome
}

