% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fagin.R
\name{load_gff}
\alias{load_gff}
\title{Load a species GFF into the database}
\usage{
load_gff(name, gff)
}
\arguments{
\item{name}{species name}

\item{gff}{gene feature format file}
}
\description{
Load a species GFF into the database
}

