% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{load_proteome}
\alias{load_proteome}
\title{Load a species predicted proteins}
\usage{
load_proteome(name, proteome)
}
\arguments{
\item{name}{the species scientific name}

\item{proteome}{a proteomic fasta file name}
}
\description{
Load a species predicted proteins
}

