% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{load_species_tree}
\alias{load_species_tree}
\title{Load the species tree}
\usage{
load_species_tree(treefile)
}
\arguments{
\item{treefile}{a Newick format tree file}
}
\value{
tree species tree
}
\description{
Load the species tree
}

