% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{load_tree_io}
\alias{load_tree_io}
\title{Load a nexus format tree}
\usage{
load_tree_io(treefile)
}
\arguments{
\item{treefile}{A filename for a nexus tree}
}
\value{
a phylo object
}
\description{
Load a nexus format tree
}
