% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{make_excel_spreadsheet}
\alias{make_excel_spreadsheet}
\title{Create an excel spreadsheet of fagin results}
\usage{
make_excel_spreadsheet(m, filename = "fagin-result.xlsx")
}
\arguments{
\item{con}{Config object}
}
\description{
Contains all the tabular data created by the \code{make_result_archive} function.
}
