% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cache.R
\name{make_fagin_cacher}
\alias{make_fagin_cacher}
\title{Build the cacher that will be used inside rmonad}
\usage{
make_fagin_cacher()
}
\value{
rmonad::Cacher object
}
\description{
Build the cacher that will be used inside rmonad
}
