% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmetics_warnings.R
\name{make_format_translation_warning}
\alias{make_format_translation_warning}
\title{Collate the warnings of a Biostrings::translate function}
\usage{
make_format_translation_warning(label = NULL)
}
\arguments{
\item{label}{The species name, usually}

\item{node}{Rmonad object wrapping the result of a Biostrings::translate function}
}
\value{
Rmonad An Rmonad object with a modified warning field
}
\description{
The Biostrings::translate function emits one warning for each CDS that is
not a multiple of 3. This function collates all these warnings into one.
Also it specifies the name of each sequence and the species from whch they
derive.
}
\details{
The warning is assumed to have the form:
"in 'x[[123]]': last 2 bases were ignored"
"in 'x[[456]]': last base was ignored"
}
