% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{make_genome_table}
\alias{make_genome_table}
\title{Tabulate states for each genome}
\usage{
make_genome_table(m, species_order)
}
\arguments{
\item{m}{Rmonad}

\item{species_order}{all species in tree order}
}
\description{
Tabulate states for each genome
}
