% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{make_origin_table}
\alias{make_origin_table}
\title{Make a table of the origin summaries for each gene}
\usage{
make_origin_table(m)
}
\arguments{
\item{m}{Rmonad object}
}
\description{
Make a table of the origin summaries for each gene
}
