% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{make_query_target_table}
\alias{make_query_target_table}
\title{Tabulate the query to target relationships}
\usage{
make_query_target_table(m)
}
\arguments{
\item{m}{Rmonad object storing the results of a successful run}
}
\value{
data.frame results
}
\description{
This function takes the final Rmonad object returned from a successful fagin
run and tabulates the binary features, secondary classes, and primary
classes for each focal gene against each target species.
}
