% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{make_secondary_genewise_table}
\alias{make_secondary_genewise_table}
\title{Make a table of the secondary labels for each query gene}
\usage{
make_secondary_genewise_table(m)
}
\arguments{
\item{m}{Rmonad object}
}
\description{
Make a table of the secondary labels for each query gene
}
