% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{make_secondary_labels_table}
\alias{make_secondary_labels_table}
\title{Summarize secondary labels as a table}
\usage{
make_secondary_labels_table(m, species_order)
}
\arguments{
\item{m}{Rmonad}

\item{species_order}{all target species in tree order}
}
\description{
Summarize secondary labels as a table
}
