% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{make_synder_flag_table}
\alias{make_synder_flag_table}
\title{Summarize the synder flags for each species}
\usage{
make_synder_flag_table(m, group = "query")
}
\arguments{
\item{m}{Romand object}

\item{group}{string - either "query" or "control"}
}
\description{
Summarize the synder flags for each species
}
