% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{make_synder_table}
\alias{make_synder_table}
\title{Tabulate states for each genome}
\usage{
make_synder_table(m)
}
\arguments{
\item{m}{Rmonad object}

\item{group}{string - either "query" or "control"}
}
\description{
Tabulate states for each genome
}
