% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{makeExcelSpreadsheet}
\alias{makeExcelSpreadsheet}
\title{Create an excel spreadsheet of fagin results}
\usage{
makeExcelSpreadsheet(con, filename = "fagin-result.xlsx")
}
\arguments{
\item{con}{Config object}
}
\description{
Contains all the tabular data created by the \code{makeResultArchive} function.
}
