% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{makeGenomeTable}
\alias{makeGenomeTable}
\title{Tabulate states for each genome}
\usage{
makeGenomeTable(m, speciesOrder)
}
\arguments{
\item{m}{Rmonad}

\item{speciesOrder}{all species in tree order}
}
\description{
Tabulate states for each genome
}
