% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{makeOriginTable}
\alias{makeOriginTable}
\title{Make a table of the origin summaries for each gene}
\usage{
makeOriginTable(m)
}
\arguments{
\item{m}{Rmonad object}
}
\description{
Make a table of the origin summaries for each gene
}
