% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{makeResultArchive}
\alias{makeResultArchive}
\title{Organize output data into a small archive}
\usage{
makeResultArchive(con)
}
\arguments{
\item{con}{Config object}
}
\value{
filename of the output archive
}
\description{
Running \code{run_fagin} will produce a large archive with cached data and
many intermediate files. This function digests that archive into a clean,
succinct, well-organized folder.
}
