% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{makeSecondaryTable}
\alias{makeSecondaryTable}
\title{Make a table of the secondary labels for each query gene}
\usage{
makeSecondaryTable(m)
}
\arguments{
\item{m}{Rmonad object}
}
\description{
Make a table of the secondary labels for each query gene
}
