% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{makeSynderFlagTable}
\alias{makeSynderFlagTable}
\title{Summarize the synder flags for each species}
\usage{
makeSynderFlagTable(m, group = "query")
}
\arguments{
\item{m}{Romand object}

\item{group}{string - either "query" or "control"}
}
\description{
Summarize the synder flags for each species
}
