% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{makeSynderTable}
\alias{makeSynderTable}
\title{Tabulate states for each genome}
\usage{
makeSynderTable(con, speciesOrder)
}
\arguments{
\item{con}{Config object}
}
\description{
Tabulate states for each genome
}
