% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{makeSynmapTable}
\alias{makeSynmapTable}
\title{Tabulate states for each genome}
\usage{
makeSynmapTable(con, speciesOrder)
}
\arguments{
\item{con}{Config object}
}
\description{
Tabulate states for each genome
}
