% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_misc.R
\name{merge_named_vectors}
\alias{merge_named_vectors}
\title{Merge named vectors into a data.frame}
\usage{
merge_named_vectors(.fill = NA, ...)
}
\arguments{
\item{.fill}{value to replace NAs with}

\item{...}{named vectors}
}
\description{
If the input list of vectors has names, these names will be used as the
column names. Otherwise automatically generated names will be used (e.g.
'value.x').
}
