% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{number_of_chromosomes}
\alias{number_of_chromosomes}
\title{Get a table listing the scaffold count for each species}
\usage{
number_of_chromosomes(x)
}
\description{
Get a table listing the scaffold count for each species
}
