% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{number_of_scaffolds}
\alias{number_of_scaffolds}
\title{Get a table listing the scaffold count for each species}
\usage{
number_of_scaffolds(m)
}
\arguments{
\item{m}{Rmonad}
}
\description{
Get a table listing the scaffold count for each species
}
