% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{parse_feature_graph}
\alias{parse_feature_graph}
\title{Parse parent/child relationships from GFF file}
\usage{
parse_feature_graph(x)
}
\arguments{
\item{x}{GFF}
}
\value{
[(Character, Character)]
}
\description{
Parse parent/child relationships from GFF file
}

