% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{plotSecondaryLabels}
\alias{plotSecondaryLabels}
\title{Plot the secondary labels}
\usage{
plotSecondaryLabels(m, speciesOrder, fill = "secondary")
}
\arguments{
\item{m}{Rmonad}

\item{speciesOrder}{all target species in tree order}

\item{fill}{character - either 'secondary' or not}
}
\description{
Plot the secondary labels
}
