% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{primary_data}
\alias{primary_data}
\title{Derive secondary data from the minimal required inputs}
\usage{
primary_data(...)
}
\arguments{
\item{con}{The config object that provides paths to required data}
}
\value{
derived_input object
}
\description{
Derive secondary data from the minimal required inputs
}
