% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{protein_has_internal_stop}
\alias{protein_has_internal_stop}
\title{Count the internal stop codons in each species protein models}
\usage{
protein_has_internal_stop(x)
}
\description{
Count the internal stop codons in each species protein models
}
