% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fagin.R
\name{run_fagin}
\alias{run_fagin}
\title{Run a Fagin analysis}
\usage{
run_fagin(con)
}
\arguments{
\item{con}{A fagin_config object}
}
\value{
An rmonad object containing all results
}
\description{
Run a Fagin analysis
}
