% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_sequence_alignments.R
\name{seqFromGenomicRange}
\alias{seqFromGenomicRange}
\title{Extract a DNAStringSet from a DNAStringSet and a GRange object}
\usage{
seqFromGenomicRange(gr, fa, scramble = FALSE)
}
\arguments{
\item{gr}{A GRange object}

\item{fa}{A DNAStringSet object}
}
\value{
DNAStringSet
}
\description{
Extract a DNAStringSet from a DNAStringSet and a GRange object
}
