% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{synmap_summary-class}
\alias{synmap_summary-class}
\alias{synmap_summary}
\title{Summary of a single synteny map}
\description{
Summary of a single synteny map
}
\section{Slots}{

\describe{
\item{\code{nrow}}{integer}

\item{\code{query_target_length_ratio}}{numeric_summary}

\item{\code{length_summary}}{numeric_summary}

\item{\code{score_summary}}{numeric_summary}

\item{\code{length_to_score_coref}}{matrix}
}}

