% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{validator}
\alias{validate_aa_aln_result}
\alias{validate_faa}
\alias{validate_feature_matrix}
\alias{validate_fna}
\alias{validate_frn}
\alias{validate_gff}
\alias{validate_gff_fna}
\alias{validate_origins}
\alias{validate_query_list}
\alias{validate_species_tree}
\alias{validate_synder}
\alias{validate_synmap}
\alias{validate_synmap_synder}
\alias{validate_trees_labels_summary}
\alias{validator}
\title{Data types and validators}
\usage{
validate_synmap(x)

validate_synder(x)

validate_gff(x)

validate_fna(x)

validate_frn(x)

validate_faa(x)

validate_species_tree(x)

validate_query_list(x)

validate_aa_aln_result(x)

validate_trees_labels_summary(x)

validate_origins(x)

validate_feature_matrix(x)

validate_synmap_synder(x)

validate_gff_fna(x)
}
\arguments{
\item{x}{data to be checked}
}
\description{
Simple types
\itemize{
  \item Bool
  \item Character
  \item Numeric
  \item Integer
  \item Count
}
}
\details{
Compound types
\itemize{
  \item Synmap
  \item SearchInterval
  \item GFF
  \item Genome
  \item CDS
  \item Protein
  \item Phylogeny
  \item GeneID
  \item ProtMSA
  \item NuclMSA
}
}

