% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeast_config.R
\name{yeast_config}
\alias{yeast_config}
\title{Initialize the configuration for the yeast case study}
\usage{
yeast_config()
}
\value{
Config object for yeast using data included in the package
}
\description{
Initialize the configuration for the yeast case study
}
