% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_aa}
\alias{derive_aa}
\title{Derive protein sequence from genome and gene model GFF}
\usage{
derive_aa(gff, dna)
}
\arguments{
\item{gff}{GenomicRanges object annotated with \code{type} and \code{parent}}

\item{dna}{DNAStringSet genome}
}
\value{
AAStringSet object
}
\description{
Derive protein sequence from genome and gene model GFF
}
