% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_orffaa}
\alias{derive_orffaa}
\title{Derive ORF protein sequences from genome and ORF ranges}
\usage{
derive_orffaa(dna, orfgff)
}
\arguments{
\item{dna}{DNAStringSet genome}

\item{orfgff}{GenomicRanges object}
}
\value{
AAStringSet object
}
\description{
Derive ORF protein sequences from genome and ORF ranges
}
