% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_orfgff}
\alias{derive_orfgff}
\title{Derive mono-exonic ORF intervals from a genome}
\usage{
derive_orfgff(dna)
}
\arguments{
\item{dna}{DNAStringSet genome}
}
\value{
GenomicRanges object
}
\description{
Note, here I am doing the stupidest thing possible: just finding all
'<START>(...)+<STOP>' intervals. I should replace this with a real ORF
finder.
}
