% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive.R
\name{derive_trans}
\alias{derive_trans}
\title{Derive mRNA sequence from genome and gene model GFF}
\usage{
derive_trans(dna, gff)
}
\arguments{
\item{dna}{DNAStringSet genome}

\item{gff}{GenomicRanges object annotated with \code{type} and \code{parent}}
}
\value{
DNAStringSet object
}
\description{
Derive mRNA sequence from genome and gene model GFF
}
