% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_4.R
\name{determine_origins}
\alias{determine_origins}
\title{Given a species tree and a set of labels, determine orphan origin}
\usage{
determine_origins(labels, con)
}
\description{
Given a species tree and a set of labels, determine orphan origin
}
