% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origins.R
\name{determineOrigins}
\alias{determineOrigins}
\title{Given a species tree and a set of labels, determine orphan origin}
\usage{
determineOrigins(labels, con)
}
\description{
Given a species tree and a set of labels, determine orphan origin
}
