% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_cache.R
\name{fagin_cache}
\alias{fagin_cache}
\alias{to_cache}
\alias{from_cache}
\title{Store and retrieve data using groups and labels}
\usage{
to_cache(x, group, label)

from_cache(group, label)
}
\arguments{
\item{x}{Data to be cached}

\item{group}{a caching group (e.g. a species name)}

\item{label}{the base label under which to store the data}
}
\description{
Currently I am caching by saving the files in as RData objects in a local
folder. However, I may replace this at some point with a database. This API
should be general enough to accomadate such changes. Hence I refer to the
data 'group', rather than directory and 'label' rather than basename.
}
