% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cache.R
\name{fagin_cache}
\alias{fagin_cache}
\alias{make_cache_function}
\alias{from_cache}
\title{Store and retrieve data using groups and labels}
\usage{
make_cache_function(cache_dir)

from_cache(file, type = "Rdata")
}
\arguments{
\item{cache_dir}{The directory in which to cache the results}

\item{x}{Data to be cached}

\item{group}{A caching group (e.g. a species name)}

\item{label}{The base label under which to store the data}
}
\description{
Currently I am caching by saving the files in as RData objects in a local
folder. However, I may replace this at some point with a database. This API
should be general enough to accomadate such changes. Hence I refer to the
data 'group', rather than directory and 'label' rather than basename.
}
