% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{genomic_composition}
\alias{genomic_composition}
\title{Summarize the genomic composition of each species}
\usage{
genomic_composition(x)
}
\description{
Summarize the genomic composition of each species
}
