% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_handling.R
\name{load_gene_models}
\alias{load_gene_models}
\title{Load a GFF file as a GenomicFeatures object}
\usage{
load_gene_models(file, seqinfo_ = NULL)
}
\arguments{
\item{file}{GFF filename}

\item{seqinfo_}{Seqinfo object}

\item{Rmonad}{wrapped GenomicFeatures object}
}
\description{
Load a GFF file as a GenomicFeatures object
}
