% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fagin.R
\name{load_genome}
\alias{load_genome}
\title{Load a species GFF into the database}
\usage{
load_genome(name, genome)
}
\arguments{
\item{name}{species name}

\item{genome}{file name}
}
\description{
Load a species GFF into the database
}

