% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gff_handling.R
\name{load_gff}
\alias{load_gff}
\title{Load a GFF file as a GenomicFeatures object}
\usage{
load_gff(file, tags, get_naked = FALSE, infer_id = FALSE, seqinfo_ = NULL)
}
\arguments{
\item{file}{GFF filename}

\item{tags}{metadata tags to keep}

\item{get_naked}{parse untagged attributes as ID if they are they only field}

\item{infer_id}{if ID is missing, try to create one}

\item{Rmonad}{wrapped GenomicFeatures object}
}
\description{
Load a GFF file as a GenomicFeatures object
}
