% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{load_synmap}
\alias{load_synmap}
\title{Load a synteny map}
\usage{
load_synmap(qname, tname, synmap)
}
\arguments{
\item{qname}{name of query species}

\item{tname}{name of target species}

\item{synmap}{synteny map file name}
}
\description{
Load a synteny map
}

