% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_cache.R
\name{make_fagin_cacher}
\alias{make_fagin_cacher}
\title{Build the cacher that will be used inside rmonad}
\usage{
make_fagin_cacher(archive_dir, cache_dir)
}
\arguments{
\item{archive_dir}{The main archive directory}

\item{cache_dir}{The rmonad cache directory (relative to the archive directory)}
}
\description{
Build the cacher that will be used inside rmonad
}
