% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{makeGenomeTable}
\alias{makeGenomeTable}
\title{Tabulate states for each genome}
\usage{
makeGenomeTable(m, speciesOrder)
}
\arguments{
\item{list}{Summarized data as output by \code{invertSummaries}}
}
\description{
Tabulate states for each genome
}
