% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploration.R
\name{plot_secondary_labels}
\alias{plot_secondary_labels}
\title{Plot the secondary labels}
\usage{
plot_secondary_labels(m, species_order, strata, fill = "secondary")
}
\arguments{
\item{m}{Rmonad}

\item{species_order}{all target species in tree order}

\item{strata}{A table with the phylostrata of each query gene}

\item{fill}{character - either 'secondary' or not}
}
\description{
Plot the secondary labels
}
