% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{seq_summary-class}
\alias{seq_summary-class}
\alias{seq_summary}
\title{Summary of a sequence}
\description{
This is a parent to both DNA and protien summary classes
}
\section{Slots}{

\describe{
\item{\code{seqids}}{character the name of each entry}

\item{\code{sizes}}{integer lengths of each sequence}

\item{\code{nseq}}{integer total number of sequences}

\item{\code{comp}}{matrix base composition}

\item{\code{comp_dist}}{list}

\item{\code{n_non_canonical}}{integer}

\item{\code{length_sum}}{numeric_summary}
}}

